requirements.txt
# Python version requirement
# Requires Python 3.6+ for f-string support

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0
scipy>=1.7.0

# Data processing and visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
tqdm>=4.62.0

# Machine learning and deep learning
torch>=1.9.0
torchvision>=0.10.0
albumentations>=1.1.0
imgaug>=0.4.0

# Configuration and utilities
PyYAML>=6.0
click>=8.0.0
pathlib2>=2.3.0
shutil
glob2>=0.7

# Image processing and augmentation
imageio>=2.9.0
scikit-learn>=1.0.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Progress tracking and logging
rich>=10.0.0
loguru>=0.5.0

# Data validation and quality
jsonschema>=3.2.0
pydantic>=1.8.0
